Move env variables in a better place
