+++ jest.config.js
@@ -2,5 +2,5 @@ module.exports = {
   roots: ['<rootDir>/src'],
   testMatch: ['**/*.test.ts'],
   collectCoverage: true,
-  coveragePathIgnorePatterns: ['ormconfig.ts', 'src/config'],
+  coveragePathIgnorePatterns: ['ormconfig.ts', 'src/envs.ts', 'src/config'],
 }
