+++ src/server.ts
@@ -2,7 +2,7 @@ import app from './app'
 import { logger } from './lib'
 
 import { databaseConnection } from './config'
-import { appUrl, port } from './dotenv'
+import { appUrl, port } from './envs'
 
 databaseConnection.create().then(() => {
   app.listen(port, () => {
